requirements.txt
tensorflow
numpy
Pillow
requests
selenium